Improvement of registration and login form
* added crispy in form
* created forms file
* added bootstrap buttons
